+++ A02_air_track_framework.py
@@ -124,16 +124,16 @@ class AirTrack:
 
 class Environment:
     def __init__(self, length_px, length_m):
-        self.px_to_m = length_m/float(length_px)
-        self.m_to_px = (float(length_px)/length_m)
+        self.px_to_m = length_m/length_px
+        self.m_to_px = length_px/length_m
     
     # Convert from meters to pixels
     def px_from_m(self, dx_m):
-        return int(round(dx_m * self.m_to_px))
+        return round(dx_m * self.m_to_px)
     
     # Convert from pixels to meters
     def m_from_px(self, dx_px):
-        return float(dx_px) * self.px_to_m
+        return dx_px * self.px_to_m
         
     def get_local_user_input(self):
         
@@ -209,7 +209,7 @@ def main():
         game_window.surface.fill(THECOLORS["black"])
 
         # Get the delta t for one frame (this changes depending on system load).
-        dt_s = float(myclock.tick(framerate_limit) * 1e-3)
+        dt_s = myclock.tick(framerate_limit) * 1e-3
         
         # Check for user initiated stop or demo change.
         resetmode = env.get_local_user_input()
