+++ A06_air_track_gui_controls.py
@@ -153,7 +153,7 @@ class Detroit:
         self.width_px  = width_px
         
         # Use y midpoint for drawing the cursor line.
-        self.center_y_px = int(round( float(game_window.height_px - self.height_px) + float(self.height_px)/2.0) )
+        self.center_y_px = round((game_window.height_px - self.height_px) + self.height_px/2.0)
         # For use with cursor-tethers selection.
         self.selected = False
         
@@ -371,8 +371,8 @@ class AirTrack:
 
 class Environment:
     def __init__(self, length_px, length_m):
-        self.px_to_m = length_m/float(length_px)
-        self.m_to_px = (float(length_px)/length_m)
+        self.px_to_m = length_m/length_px
+        self.m_to_px = length_px/length_m
         
         # Add a local (non-network) client to the client dictionary.
         self.clients = {'local':Client(THECOLORS["green"])}
@@ -380,11 +380,11 @@ class Environment:
     
     # Convert from meters to pixels
     def px_from_m(self, dx_m):
-        return int(round(dx_m * self.m_to_px))
+        return round(dx_m * self.m_to_px)
     
     # Convert from pixels to meters
     def m_from_px(self, dx_px):
-        return float(dx_px) * self.px_to_m
+        return dx_px * self.px_to_m
         
     def get_local_user_input(self):
         
@@ -396,7 +396,7 @@ class Environment:
                 if (event.key == K_ESCAPE):
                     return 'quit'
                 elif (event.key==K_1):            
-                    return 1           
+                    return 1
                 elif (event.key==K_2):                          
                     return 2
                 elif (event.key==K_3):                          
@@ -506,7 +506,7 @@ def main():
         game_window.surface.fill(THECOLORS["black"])
 
         # Get the delta t for one frame (this changes depending on system load).
-        dt_s = float(myclock.tick(framerate_limit) * 1e-3)
+        dt_s = myclock.tick(framerate_limit) * 1e-3
 
         # This check avoids problem when dragging the game window.
         if (dt_s < 0.10):
